# [plugin][mercurial] Optimize in_hg and hg_get_branch_name