When not in the repo root dir `in_hg` used `hg summary` which is terribly slow on big repositories, and `hg_get_branch_name` called `hg branch` which is also pretty slow as well.

Based on the [branch](https://github.com/robbyrussell/oh-my-zsh/tree/master/plugins/branch) plugin, I created the function `hg_get_dir` that fetch the `.hg` directory if it exists, going up in the directory tree. From there, we know if we're in an HG repository if `in_hg` returns a non empty string, and we can simply get the branch name from the content of the `.hg/branch` file.

### Benchmarks

Based on 10 calls to `hg_get_branch_name` in subdir (so that the old `hg summary` is called), here are the average times:
* Before: `/tmp/test.zsh  0.08s user 0.04s system 96% cpu 0.133 total`
* After: `/tmp/test.zsh  0.00s user 0.00s system 71% cpu 0.008 total`

It's faster and it uses less CPU!

The other changes after that are purely esthetic (fit in 80 columns and indentation fix).