@MightyCreak 2018-02-18
You are right for both, I don't even have that in my local file, I don't know how these errors got there... noob mistake I guess ;)

I'll upload a new PR