@MightyCreak 2018-02-18
Ok, there's been too much changes since 2016.
Let me do another proper PR.